+++ collector/mongodb_collector.go
@@ -15,8 +15,8 @@
 package collector
 
 import (
-	"errors"
 	"fmt"
+	"sync"
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
@@ -40,6 +40,7 @@ type MongodbCollectorOpts struct {
 	TLSHostnameValidation bool
 	CollectDatabaseMetrics   bool
 	CollectCollectionMetrics bool
+	DBPoolLimit           int
 }
 
 func (in MongodbCollectorOpts) toSessionOps() shared.MongoSessionOpts {
@@ -50,15 +51,20 @@ func (in MongodbCollectorOpts) toSessionOps() shared.MongoSessionOpts {
 		TLSPrivateKeyFile:     in.TLSPrivateKeyFile,
 		TLSCaFile:             in.TLSCaFile,
 		TLSHostnameValidation: in.TLSHostnameValidation,
+		PoolLimit:             in.DBPoolLimit,
 	}
 }
 
 // MongodbCollector is in charge of collecting mongodb's metrics.
 type MongodbCollector struct {
-	Opts                      MongodbCollectorOpts
+	Opts MongodbCollectorOpts
+
 	scrapesTotal              prometheus.Counter
 	lastScrapeError           prometheus.Gauge
 	lastScrapeDurationSeconds prometheus.Gauge
+
+	mongoSessLock sync.Mutex
+	mongoSess     *mgo.Session
 }
 
 // NewMongodbCollector returns a new instance of a MongodbCollector.
@@ -87,6 +93,31 @@ func NewMongodbCollector(opts MongodbCollectorOpts) *MongodbCollector {
 	}
 
 	return exporter
+}
+
+// getSession returns the cached *mgo.Session or creates a new session and returns it.
+// Use sync.Mutex to avoid race condition around session creation.
+func (exporter *MongodbCollector) getSession() *mgo.Session {
+	exporter.mongoSessLock.Lock()
+	defer exporter.mongoSessLock.Unlock()
+
+	if exporter.mongoSess == nil {
+		exporter.mongoSess = shared.MongoSession(exporter.Opts.toSessionOps())
+	}
+	if exporter.mongoSess == nil {
+		return nil
+	}
+	return exporter.mongoSess.Copy()
+}
+
+// Close cleanly closes the mongo session if it exists.
+func (exporter *MongodbCollector) Close() {
+	exporter.mongoSessLock.Lock()
+	defer exporter.mongoSessLock.Unlock()
+
+	if exporter.mongoSess != nil {
+		exporter.mongoSess.Close()
+	}
 }
 
 // Describe sends the super-set of all possible descriptors of metrics collected by this Collector
@@ -141,9 +172,10 @@ func (exporter *MongodbCollector) scrape(ch chan<- prometheus.Metric) {
 		}
 	}(time.Now())
 
-	mongoSess := shared.MongoSession(exporter.Opts.toSessionOps())
+	mongoSess := exporter.getSession()
 	if mongoSess == nil {
-		err = errors.New("can't create mongo session")
+		err = fmt.Errorf("Can't create mongo session to %s", exporter.Opts.URI)
+		log.Error(err)
 		return
 	}
 	defer mongoSess.Close()