+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.4.0 (not released yet)
+
+* Go vendoring switched to [dep](https://github.com/golang/dep).
+
+## v0.3.1 (2017-09-08)
+
+* Better logging for scrape errors.
+
 ## v0.3.0 (2017-07-07)
 
 * Add standard metrics: